# Add BoundedArray and use it to avoid allocating vectors