--- a/base/base.vcxproj
+++ b/base/base.vcxproj
@@ -302,6 +302,7 @@
     <ClInclude Include="version.hpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="array_test.cpp" />
     <ClCompile Include="bundle.cpp" />
     <ClCompile Include="bundle_test.cpp" />
     <ClCompile Include="disjoint_sets_test.cpp" />