Before:
```
Benchmark                                            Time           CPU Iterations
----------------------------------------------------------------------------------
BM_VisibleSegmentsOrbit/10                        2014 ns       1943 ns     345165 average visible segments: 1.000000
BM_VisibleSegmentsOrbit/100                      12250 ns      12238 ns      56089 average visible segments: 0.980000
BM_VisibleSegmentsOrbit/1000                    116502 ns     114385 ns       6410 average visible segments: 0.966000
BM_VisibleSegmentsRandomEverywhere/1000         128759 ns     127938 ns       5609 average visible segments: 1.073000
BM_VisibleSegmentsRandomNoIntersection/1000      93610 ns      93863 ns       7479 average visible segments: 1.000000
```

After:
```
Benchmark                                            Time           CPU Iterations
----------------------------------------------------------------------------------
BM_VisibleSegmentsOrbit/10                        1196 ns       1196 ns     560894 average visible segments: 1.000000
BM_VisibleSegmentsOrbit/100                       6736 ns       6606 ns      89743 average visible segments: 0.980000
BM_VisibleSegmentsOrbit/1000                     63809 ns      63969 ns      11218 average visible segments: 0.966000
BM_VisibleSegmentsRandomEverywhere/1000          68781 ns      68141 ns      11218 average visible segments: 1.073000
BM_VisibleSegmentsRandomNoIntersection/1000      44980 ns      44778 ns      16026 average visible segments: 1.000000
```

Another contribution to #973.